check annonce and image links

+++ app/src/Resource/AnnonceResource.php
@@ -2,6 +2,7 @@
 namespace App\Resource;
 
 use App\AbstractResource;
+use App\Entity\Annonce;
 
 /**
  * Class Resource
@@ -63,12 +64,61 @@ class AnnonceResource extends AbstractResource
      *
      * @return boolean
      */
-    public function update($annonce)
-{
-    $this->entityManager->persist($annonce);
-    $this->entityManager->flush();
-    return true;
-}
+    public function insert($annonce)
+    {
+        $this->entityManager->persist($annonce);
+        $this->entityManager->flush();
+        return true;
+    }
+    /**
+     * @param object $annonce
+     *
+     * @return boolean
+     */
+    public function update($post,$key)
+
+    {
+
+        $tmp = new Annonce();
+
+        if ( $key!== null)
+        $tmp = $this->entityManager->getRepository('App\Entity\Annonce')->findOneBy($key);
+
+
+        if ($post->getAnnoncestate()!==null){
+            $tmp->setAnnoncestate($post->getAnnoncestate());
+        }
+        if ( $post->getCoords()!==null){
+            $tmp->setCoords($post->getCoords());
+        }
+        if ( $post->getTitre()!==null){
+            $tmp->setTitre($post->getTitre());
+        }
+        if ( $post->getType()!==null){
+            $tmp->setCity($post->getType());
+        }
+        if ( $post->getCity()!==null){
+            $tmp->setTitre($post->getCity());
+        }
+        if ( $post->getRooms()!==null){
+            $tmp->setRooms($post->getRooms());
+        }
+        if ( $post->getPrixtotal()!==null){
+            $tmp->setPrixtotal($post->getPrixtotal());
+        }
+        if ( $post->getTitre()!==null){
+            $tmp->setTitre($post->getTitre());
+        }
+
+        $this->entityManager->persist($tmp);
+        $this->entityManager->flush();
+        return true;
+    }
+
+
+
+
+
 
     /**
      * @param String $key
@@ -77,12 +127,16 @@ class AnnonceResource extends AbstractResource
      */
     public function delete($key)
     {
-        $annonce = $this->entityManager->getRepository('App\Entity\Annonce')->findOneBy($key);
-
-        $this->entityManager->remove($annonce);
+        $annonces = $this->entityManager->getRepository('App\Entity\Annonce')->findBy($key);
+    if (!($annonces=== null) ){
+        foreach ($annonces as $annonce) {
+            $this->entityManager->remove($annonce);
+        }
         $this->entityManager->flush();
-
         return true;
-
     }
+        return false;
+}
+
+
 }
